Apply patch to fix some compiler warnings and related issues
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/8578


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@30205 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/locmap.c
@@ -957,7 +957,7 @@ uprv_convertToPosix(uint32_t hostid, UErrorCode* status)
 
     ret = GetLocaleInfoA(hostid, LOCALE_SNAME, (LPSTR)gPosixFromLCID, sizeof(gPosixFromLCID));
     if (ret > 1) {
-        FIX_LOCALE_ID_TAG_SEPARATOR(gPosixFromLCID, ret, localeIndex)
+        FIX_LOCALE_ID_TAG_SEPARATOR(gPosixFromLCID, (uint32_t)ret, localeIndex)
         FIX_LANGUAGE_ID_TAG(gPosixFromLCID, ret)
 
         return gPosixFromLCID;
